tweak: address code review
